Week2/ec2-layer/django-app/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
drf-yasg
